+++ src/com/intellij/idea/plugin/hybris/spring/context/CngSpringELContextsExtension.kt
@@ -19,18 +19,24 @@ package com.intellij.idea.plugin.hybris.spring.context
 
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.project.utils.PluginCommon
-import com.intellij.idea.plugin.hybris.system.cockpitng.CngConfigDomFileDescription
+import com.intellij.idea.plugin.hybris.system.cockpitng.CngConfigDomFileDescription.Companion.NAMESPACE_COCKPIT_NG_CONFIG_HYBRIS
+import com.intellij.idea.plugin.hybris.system.cockpitng.CngConfigDomFileDescription.Companion.NAMESPACE_COCKPIT_NG_CONFIG_WIZARD_CONFIG
 import com.intellij.idea.plugin.hybris.system.cockpitng.model.config.hybris.Labels
+import com.intellij.idea.plugin.hybris.system.cockpitng.model.wizardConfig.AbstractAction
 import com.intellij.idea.plugin.hybris.system.cockpitng.psi.CngPsiHelper
 import com.intellij.javaee.el.util.ELImplicitVariable
 import com.intellij.openapi.project.Project
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiTypes
-import com.intellij.psi.PsiVariable
+import com.intellij.psi.*
+import com.intellij.psi.impl.light.DefiniteLightVariable
 import com.intellij.psi.scope.processor.MethodResolveProcessor
 import com.intellij.psi.search.GlobalSearchScope
 import com.intellij.psi.search.PsiShortNamesCache
 import com.intellij.psi.util.PropertyUtilBase
+import com.intellij.psi.util.PsiTypesUtil
+import com.intellij.psi.util.parentOfType
+import com.intellij.psi.xml.XmlAttribute
+import com.intellij.psi.xml.XmlAttributeValue
+import com.intellij.psi.xml.XmlTag
 import com.intellij.psi.xml.XmlText
 import com.intellij.spring.el.contextProviders.SpringElContextsExtension
 
@@ -42,35 +48,61 @@ class CngSpringELContextsExtension : SpringElContextsExtension() {
         val context = spelFile.context ?: return mutableListOf()
         val project = spelFile.project
 
-        if (context is XmlText) {
-            val tag = context.parentTag ?: return mutableListOf()
+        val tag = context.parentOfType<XmlTag>() ?: return mutableListOf()
 
-            if (tag.localName == Labels.LABEL && tag.namespace == CngConfigDomFileDescription.NAMESPACE_COCKPIT_NG_CONFIG_HYBRIS) {
-                return getContextVariablesCngLabel(context, project)
-            }
+        return when {
+            context is XmlText
+                && tag.localName == Labels.LABEL
+                && tag.namespace == NAMESPACE_COCKPIT_NG_CONFIG_HYBRIS -> process(context, project)
+
+            context is XmlAttributeValue
+                && context.parentOfType<XmlAttribute>()?.localName == AbstractAction.VISIBLE
+                && tag.namespace == NAMESPACE_COCKPIT_NG_CONFIG_WIZARD_CONFIG -> process(context, project)
+
+            else -> mutableListOf()
         }
-        return mutableListOf()
     }
 
-    private fun getContextVariablesCngLabel(context: XmlText, project: Project) = context
+    private fun process(context: XmlText, project: Project) = context
         .let { CngPsiHelper.resolveContextType(it) }
+        ?.let { findClassByHybrisTypeName(project, it) }
+        ?.let { process(it) }
+        ?.toMutableList()
+        ?: mutableListOf()
+
+    private fun process(context: XmlAttributeValue, project: Project) = context
+        .let { CngPsiHelper.resolveNamedContextTypeForNewItemInWizardFlow(it) }
         ?.let {
-            PsiShortNamesCache.getInstance(project).getClassesByName(
-                it + HybrisConstants.MODEL_SUFFIX, GlobalSearchScope.allScope(project)
-            )
-                .firstOrNull { psiClass -> psiClass.containingFile.virtualFile.path.contains("/platform/bootstrap") }
+            val name = it.first ?: return@let null
+            val type = it.second ?: return@let null
+
+            val psiClass = if (type.contains(".") && type != HybrisConstants.COCKPIT_NG_INITIALIZE_CONTEXT_TYPE) findClassByFQN(project, type)
+            else findClassByHybrisTypeName(project, type)
+            if (psiClass == null) return@let null
+
+            name to psiClass
         }
-        ?.let { MethodResolveProcessor.getAllMethods(it) }
-        ?.filter { it.hasModifierProperty("public") && !it.isConstructor }
-        ?.filter { it.returnTypeElement?.type != PsiTypes.voidType() }
-        ?.mapNotNull { method ->
+        ?.let { DefiniteLightVariable(it.first, PsiTypesUtil.getClassType(it.second), it.second) }
+        ?.let { mutableListOf(it) }
+        ?: mutableListOf()
+
+    private fun findClassByHybrisTypeName(project: Project, typeName: String) = PsiShortNamesCache.getInstance(project)
+        .getClassesByName(typeName + HybrisConstants.MODEL_SUFFIX, GlobalSearchScope.allScope(project))
+        .firstOrNull { psiClass -> psiClass.containingFile.virtualFile.path.contains("/platform/bootstrap") }
+
+    private fun findClassByFQN(project: Project, className: String) = JavaPsiFacade.getInstance(project)
+        .findClass(className, GlobalSearchScope.allScope(project))
+
+    private fun process(psiClass: PsiClass) = psiClass
+        .let { MethodResolveProcessor.getAllMethods(it) }
+        .filter { it.hasModifierProperty("public") && !it.isConstructor }
+        .filter { it.returnTypeElement?.type != PsiTypes.voidType() }
+        .mapNotNull { method ->
             val name = PropertyUtilBase.getPropertyName(method) ?: return@mapNotNull null
             val type = PropertyUtilBase.getPropertyType(method) ?: return@mapNotNull null
             val nameIdentifier = PropertyUtilBase.getPropertyNameIdentifier(method) ?: return@mapNotNull null
 
             ELImplicitVariable(method, name, type, nameIdentifier, ELImplicitVariable.NESTED_RANGE)
         }
-        ?.toMutableList()
-        ?: mutableListOf()
-
+        .toMutableList()
 }